Store interface

+++ innitdb/load.go
@@ -9,10 +9,10 @@ import (
 
 type LoadInterface interface {
 	Seed() maphash.Seed
-	Load(id uint64) (fc int, refs, irefs []float64)
+	Load(id uint64) (fc int, refs, irefs []uint64)
 }
 
-func Load(d LoadInterface, n innit.Node) (fc int, refs, irefs []float64) {
+func Load(d LoadInterface, n innit.Node) (fc int, refs, irefs []uint64) {
 	var h maphash.Hash
 	h.SetSeed(d.Seed())
 	hash.Node(&h, n)
